カートに商品を追加する機能の入力値チェック追加

+++ src/main/java/com/example/ec_201804d/repository/ItemRepository.java
@@ -97,7 +97,7 @@ public class ItemRepository {
 	 * @param id ID
 	 * @return 選択された商品情報
 	 */
-	public Item load(int id) {
+	public Item load(long id) {
 		String sql ="SELECT id,name,description,price,imagepath,deleted FROM items WHERE id=:id";
 		
 		SqlParameterSource param = new MapSqlParameterSource().addValue("id", id);
